virtio/virt_queue: change last_avail type
There is no reason to have last avail as "int".
It should be conformed with the type of (avail|used).idx which is
16 bites-wide ("_le16") by virtio specification.
Otherwise, there might be some problems with number of descriptors
difference calculation.

+++ virtio/virt_queue.h
@@ -49,7 +49,7 @@ struct virtio_virtq {
     int qsz;
 
     /* Shadow avail ring index */
-    int last_avail;
+    uint16_t  last_avail;
 
     /*
      * 2.4.5.3.1: A driver MUST NOT create a descriptor chain longer than
